requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
backcall==0.2.0
bleach==4.1.0
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.5
cycler==0.11.0
dataclasses @ file:///tmp/build/80754af9/dataclasses_1614363715916/work
datasets==2.4.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
entrypoints==0.4
filelock==3.4.1
frozenlist==1.2.0
fsspec==2022.1.0
huggingface-hub==0.4.0
idna==3.4
idna-ssl==1.1.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.2
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.0
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict==5.2.0
multiprocess==0.70.12.2
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.6
notebook==6.4.10
numpy==1.19.5
olefile==0.46
opencv-python==4.2.0.34
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.19.6
pyarrow==6.0.1
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.18.0
python-dateutil==2.8.2
pytorch-gan-metrics==0.5.2
pytz==2022.4
pywin32==304
pywinpty==1.1.6
PyYAML==6.0
pyzmq==24.0.1
requests==2.27.1
responses==0.17.0
scikit-learn==0.24.2
scipy==1.5.4
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
tensorboardX==2.5.1
terminado==0.12.1
testpath==0.6.0
threadpoolctl==3.1.0
torch==1.8.2
torchaudio==0.8.2
torchvision==0.9.2
tornado==6.1
tqdm==4.64.1
traitlets==4.3.3
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.1
wincertstore==0.2
xxhash==3.1.0
yarl==1.7.2
zipp==3.6.0
